bsy commit 0919 17:32

+++ src/main/java/com/server/cloud/s3/AwsApiController.java
@@ -75,7 +75,6 @@ public class AwsApiController {
 									String fileId){
 		Instant now = Instant.now();
 		Timestamp timestamp = Timestamp.from(now);
-		System.out.println(fileId);
 		try {
 			if(fileId!=null) {
 				
@@ -96,6 +95,7 @@ public class AwsApiController {
 			FileVO path=awsService.getImg(userId);
 			return new ResponseEntity<>(path,HttpStatus.OK);
 			}else {
+				
 				String originName=file.getOriginalFilename();
 				byte[]originData=file.getBytes();
 				String objectURI =s3.putS3Object(originName,originData);
@@ -111,6 +111,33 @@ public class AwsApiController {
 				
 				return new ResponseEntity<>("성공",HttpStatus.OK);
 			}
+		}catch (Exception e) {
+			// TODO: handle exception
+			e.printStackTrace();
+		}
+		
+		return new ResponseEntity<>("?",HttpStatus.OK);
+	}
+	@PostMapping("/api/main/cloudUploadCs")
+	public ResponseEntity<?>cloudUploadCs(@RequestParam("file_data")MultipartFile file,@RequestParam("userId")String userId,
+									String fileId){
+		Instant now = Instant.now();
+		Timestamp timestamp = Timestamp.from(now);
+		try {String originName=file.getOriginalFilename();
+				byte[]originData=file.getBytes();
+				String objectURI =s3.putS3Object(originName,originData);
+				FileVO fileVO=new FileVO().builder()
+				.file_name(originName)
+				.file_path(objectURI)
+				.file_type(file.getContentType())
+				.user_id(userId)
+				.upload_date(timestamp)
+				.build();
+				System.out.println(fileVO.toString());
+				awsService.setFileCs(fileVO);
+				
+				return new ResponseEntity<>("성공",HttpStatus.OK);
+	
 		}catch (Exception e) {
 			// TODO: handle exception
 			e.printStackTrace();
@@ -156,15 +183,29 @@ public class AwsApiController {
 	
 	@PostMapping("/api/main/AnnoDel")
 	public ResponseEntity<?>AnnoDel(@RequestBody Map<String, Object> deleteA ){
+		if(deleteA.get("file_name")!=null) {
+			System.out.println(deleteA.get("file_name"));
+			String value=(String)deleteA.get("file_name");
+			String file_num=(String)deleteA.get("file_id");
+			s3.deleteBucketObjects(value);
+			awsService.fileDel(file_num);//파일삭제
+		}
 		
+		String notice_num=(String)deleteA.get("notice_num");
+		awsService.AnnoDel(notice_num);//댓글,글 삭제
+		return null;
+	}
+	@PostMapping("/api/main/inQuryDel")
+	public ResponseEntity<?>inQuryDel(@RequestBody Map<String, Object> deleteA ){
+		if(deleteA.get("file_name")!=null) {
 		System.out.println(deleteA.get("file_name"));
 		String value=(String)deleteA.get("file_name");
 		String file_num=(String)deleteA.get("file_id");
-		String notice_num=(String)deleteA.get("notice_num");
 		s3.deleteBucketObjects(value);
-		awsService.fileDel(file_num);
-		awsService.AnnoDel(notice_num);
+		awsService.fileDel(file_num);//파일삭제
+		}
+		String notice_num=(String)deleteA.get("notice_num");
+		awsService.inQuryDel(notice_num);//댓글,글 삭제
 		return null;
 	}
-	
 }
